deprecate SolrLogPostTool in favour of the PostLogsTool
